% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_incident_uptake.R
\name{augment_with_incident}
\alias{augment_with_incident}
\title{Augment a dataframe with incident uptake}
\usage{
augment_with_incident(df)
}
\arguments{
\item{df}{Dataframe with columns \code{date}, \code{estimate}, \code{sd}}
}
\description{
Augment a dataframe with incident uptake
}
