% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_matched_method.R
\name{date_matched_method}
\alias{date_matched_method}
\title{Run uptake projection using date-matched method}
\usage{
date_matched_method(obs_uptake, ref_uptake, obs_slope)
}
\arguments{
\item{obs_uptake}{current season's uptake curve}

\item{ref_uptake}{reference uptake curve (a time series)}

\item{obs_slope}{slopes estimated for current uptake to date this season}
}
\value{
vector of cumulative uptakes
}
\description{
In the date-matched (formerly 'intermediate") method, the uptake rate continues
as it did in the reference season, based on the calendar date, regardless of the
uptake observed so far in the current season.
}
