% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_slope.R
\name{extract1}
\alias{extract1}
\title{Extract a single parameter from a Stan fit}
\usage{
extract1(x, nm)
}
\arguments{
\item{x}{Stan fit object}

\item{nm}{name of parameter}
}
\value{
vector of draws
}
\description{
Extract a single parameter from a Stan fit
}
