% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_shift.R
\name{fit_shift}
\alias{fit_shift}
\title{Least squares alignment of nis2022 and nis2023 data to find estimated delay}
\usage{
fit_shift(nis2022, nis2023)
}
\arguments{
\item{nis2022}{tibble containing nis2022 data}

\item{nis2023}{tibble containing nis2023 data}
}
\value{
shift (in days) which best aligns the two time series
}
\description{
Least squares alignment of nis2022 and nis2023 data to find estimated delay
}
