% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_incident_uptake.R
\name{get_incident_uptake}
\alias{get_incident_uptake}
\title{Compute incident uptake from cumulative}
\usage{
get_incident_uptake(x, sd)
}
\arguments{
\item{x}{cumulative uptake, ordered}

\item{sd}{standard error in \code{x}}
}
\value{
named list with incident uptake \verb{$x} and standard errors \verb{$sd}
}
\description{
Estimated standard error is an overestimate, if cumulative uptakes are positively correlated
}
