% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nis_2023.R
\name{get_nis_2023}
\alias{get_nis_2023}
\title{Get public NIS 2023 data}
\usage{
get_nis_2023(indicator = "booster")
}
\arguments{
\item{indicator}{default: \code{"booster"}, option: \code{"sentiment"}}
}
\description{
Download 2023 bivalent booster uptake data from publicly available
dataset \code{pakc-hru3}
}
\details{
The data gives the percent of people who completed their primary series, and the
percent of those people who got a bivalent booster. We compute the error in the
product of those two by assuming that the two metrics are independent, which is
a kludge.
}
