% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projection.R
\name{get_projection}
\alias{get_projection}
\title{Wrapper for projection functions}
\usage{
get_projection(name, ...)
}
\arguments{
\item{name}{character to choose method, one of "rate-matched|date-matched"}

\item{...}{Further arguments passed to the chosen method}
}
\value{
output of chosen method
}
\description{
Wrapper for projection functions
}
\seealso{
\code{\link[=rate_matched_method]{rate_matched_method()}}, \code{\link[=date_matched_method]{date_matched_method()}}
}
