% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref_uptake.R
\name{get_ref_uptake}
\alias{get_ref_uptake}
\title{Create splined reference uptake from 2022}
\usage{
get_ref_uptake(
  date,
  nis2022,
  ref_date = ymd("2023-09-15"),
  historical_ref_date = ymd("2022-09-15")
)
}
\arguments{
\item{date}{week to interpolate for}

\item{nis2022}{data input}

\item{ref_date}{2023/2024 date to use for offset}

\item{historical_ref_date}{2022/2023 season reference date}
}
\value{
uptake on a date
}
\description{
Create splined reference uptake from 2022
}
