% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_socrata.R
\name{get_socrata}
\alias{get_socrata}
\title{Get Socrata data}
\usage{
get_socrata(json_url, where = NULL, select = NULL, limit = NULL, ...)
}
\arguments{
\item{json_url}{JSON endpoint}

\item{where}{vector of filters like \code{state=AL} and \verb{state in(AL, DC)}, or a
string of filters connected by \code{" and "}}

\item{select}{vector of columns to keep, or comma-separated string of columns}

\item{limit}{number of rows (default \code{NULL} keeps all rows)}

\item{...}{additional arguments passed to \code{RSocrata::read.socrata()}}
}
\value{
\code{tibble} coerced from \code{read.socrata()}
}
\description{
Get Socrata data
}
\details{
\code{where} clauses can only use single quotes

This function constructs "clauses" that are joined into a single
URL. For example, \code{select = c("name", "age")} creates a clause
\verb{$select=name,age}. The clauses are joined, URL-encoded, and
passed through \code{read.socrata()}.
}
\examples{
# do not run by default, because of throttling limits
if (FALSE) {
  get_socrata(
    "https://healthdata.gov/resource/g62h-syeh.json",
    where = "state=AL",
    select = c("date", "state", "previous_day_admission_adult_covid_confirmed"),
    limit = 10
  )
}
}
