% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nondecreasing.R
\name{make_nondecreasing}
\alias{make_nondecreasing}
\title{Make a vector nondecreasing}
\usage{
make_nondecreasing(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
If vector is nondecreasing, do nothing. If it decreases at some step, keep the
output value the same, but accumulate a "debt". Pay down the debt with future,
positive increases.
}
