% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad}
\alias{pad}
\title{Pad a vector with a constant value}
\usage{
pad(x, len, with, side = "right")
}
\arguments{
\item{x}{Vector to pad}

\item{len}{Length to pad out to}

\item{with}{Value to pad with}

\item{side}{Side to pad onto (only \code{"right"} is supported)}
}
\value{
Vector
}
\description{
Pad a vector with a constant value
}
\examples{
pad(1:3, 5, 0)
}
