% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slopes.R
\name{plot_slopes}
\alias{plot_slopes}
\title{Plots slopes generated in fit_slope}
\usage{
plot_slopes(slope_df, data_sources = c("NIS_2022", "NIS_2023"))
}
\arguments{
\item{slope_df}{slopes data frame generated by fit_slope. includes a \code{method}
of estimation, \code{season} to group data by, \code{dataset} to group data by, and then
an \code{estimate} of the slope and an \code{lci} and \code{uci} for that estimate}

\item{data_sources}{vector of strings id-ing the data sources to be included,
need to match the sources in slope_df. Default is to not include ISD data as
they may be out of date.}
}
\value{
A ggplot object
}
\description{
Plots slopes generated in fit_slope
}
