% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propegate_sd.r
\name{product_sd}
\alias{product_sd}
\title{Propagate uncertainty}
\usage{
product_sd(mu1, mu2, sigma1, sigma2)
}
\arguments{
\item{mu1}{mean estimate of first variable}

\item{mu2}{mean estimate of second variable}

\item{sigma1}{standard deviation of first variable}

\item{sigma2}{standard deviation of second variable}
}
\description{
Propagate uncertainty
}
\details{
At multiple points in this workflow we need to propegate uncertainty across either
the product or the sum of two different estimates. We compute the error in the
product of those two by assuming that the two metrics are independent, which is
a conservative assumption.
}
