% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nis_2022.R
\name{quote_collapse}
\alias{quote_collapse}
\title{Collapse a vector of strings with single quotes and commands}
\usage{
quote_collapse(x)
}
\arguments{
\item{x}{vector of strings}
}
\value{
string
}
\description{
Collapse a vector of strings with single quotes and commands
}
