% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_matched_method.R
\name{rate_matched_method}
\alias{rate_matched_method}
\title{Run uptake projection under rate-matched method assumptions}
\usage{
rate_matched_method(
  obs_uptake,
  ref_uptake,
  obs_slope,
  use_total_slope = FALSE,
  n_observations = 4
)
}
\arguments{
\item{obs_uptake}{current season's uptake curve}

\item{ref_uptake}{reference uptake curve (a time series)}

\item{obs_slope}{slopes estimated for current uptake to date this season}

\item{use_total_slope}{option to choose slope over the full season or just
from the past \code{n} weeks for rate matching}

\item{n_observations}{to use when not using the total slope - this is the number
of weeks of data to use, starting from the most recent}
}
\value{
vector of cumulative uptakes
}
\description{
In the rate-matched (formerly "pessimistic") method, the rate of booster uptake
will decline at the same rate as it did in the reference season, so the uptake
rate is started from the matching point in that season, then the reference season
rates are used moving forward.
}
