% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_slope.R
\name{stan_slope}
\alias{stan_slope}
\title{Linear regression: fit y = mx + b line to x,y pairs to estimate the uptake slope}
\usage{
stan_slope(
  df,
  chains = 2,
  iter = 2000,
  prior_intercept_sd = 10,
  prior_slope_sd = 10,
  ...
)
}
\arguments{
\item{df}{data frame including x (week) and y (estimate) as well as sd for each data point}

\item{chains}{integer number of chains for rstan to run}

\item{iter}{integer chain length for rstan}

\item{prior_intercept_sd}{double (must be positive) see model above}

\item{prior_slope_sd}{double (must be positive) see model above}

\item{...}{further arguments to spass to \code{rstan::stan}}
}
\value{
MCMC samples as rstan::stanfit object
}
\description{
Model: y_i ~ Normal(slope * x_i + intercept, sd)
intercept ~ Normal(0, prior_intercept_sd)
slope ~ normal(0, prior_slope_sd)
sd_i := a priori estimate of standard deviation (provided by NIS)
}
