% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_slope.R
\name{stan_slope_summary}
\alias{stan_slope_summary}
\title{Summarize stan slopes}
\usage{
stan_slope_summary(df, name = "slope", ...)
}
\arguments{
\item{df}{data frame including x (week) and y (estimate) as well as sd for each data point}

\item{name}{name of the column to extract. Default: \code{"slope"}}

\item{...}{further arguments passed to \code{stan_slope()}}
}
\value{
named vector with estimate, lci, uci
}
\description{
Summarize stan slopes
}
