% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptakeprojection-package.R
\docType{package}
\name{uptakeprojection-package}
\alias{uptakeprojection}
\alias{uptakeprojection-package}
\title{uptakeprojection: COVID-19 booster uptake projections}
\description{
Given vaccine uptake data from the National Immunization Survey from the fall and early winter, make scenario-based projections about the uptake during winter and through the end of the season.
}
\author{
\strong{Maintainer}: Scott Olesen \email{ulp7@cdc.gov}

Authors:
\itemize{
  \item Inga Holmdahl \email{usn4@cdc.gov}
}

}
\keyword{internal}
