% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_slope.R
\name{bootstrap_slope}
\alias{bootstrap_slope}
\title{Nonparametric bootstrapping for fitting a line to x,y pairs to estimate the uptake slope}
\usage{
bootstrap_slope(data, n = 1000, conf = 0.95, type = "bca", min_n_rows = 3)
}
\arguments{
\item{data}{data frame including x (week) and y (estimate) as well as std for each data point}

\item{n}{number of bootstrap replicates}

\item{conf}{confidence level for bootstrap quantities (1 - alpha)}

\item{type}{passed to boot::boot.ci, type of bootstrap used (e.g. percentile, studentized, etc.)}
}
\value{
named vector; slope estimate with upper and lower 95\% confidence intervals
}
\description{
Bootstrap alternative to stan_slope() in stan_slope.R
}
